AoC 2021 - Day 24
This took a long time to work out: the spreadsheet holds most of my notes for reverse engineering the ALU code.

I first attempted to brute force a result when I went out for the day, but seven hours later it hadn't finished, so I began on the proper analysis of the code.  The result is fairly pleasing. It was somewhat of a relief when I discovered that most people were also going through similar experiences today.

Here's hoping Christmas Day follows the normal format of a fairly simple poser followed by the final star for 'free'. ;)