Library fix in grid.py
Added a new function to the grid.py library to automate the scaning
through the grid. The function is called `next_location`. It takes
the current location and the grid as arguments and returns the next
location in the grid scaning from left to right and top to bottom.
If the end of the grid is reached, the function returns None, unless
the `wrap` argument is set to True, in which case the function returns
the first location in the grid.

Additionally, the `get` and `set` functions have been modified to
accept a tuple as the location argument. This allows the efficient
use of the `next_location` function.